Realtime APIの更新ボタン追加
